pyproject.toml
[tool.poetry]
name = "chunithm課題曲bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
py-cord = "^2.6.1"
requests = "^2.32.3"
schedule = "^1.2.2"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
